2nd attempt, works with an ideal input. You can run from terminal or die, a gui will pop up that will display bot the answer and allow for an input.
Had to down tools on the first attempt, made to many security features
that meant i was moving extremely slowly. Features like non centralised
data and ambigious code (to make it harder for people to guess the code
without an end device).

Went back to the beginning and looked at how i could get something
complete, did it then spent a little extra time playing around with the
code to allow a GUI. This version took approximately six hours (closest
whole hour), a third of that was altering the code to allow the guy.